+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.3.3] - 2022-30-12
+
 ### Changed
 
 - `Messenger.transfer` to `Messenger.raw` + renaming of `MessengerTransferData` type to `MessengerRawData`